requirements.txt
youtube-transcript-api==1.2.4
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
sentence-transformers==5.3.0
faiss-cpu==1.13.2
tiktoken==0.12.0
streamlit==1.55.0
python-dotenv==1.2.2
google-ai-generativelanguage==0.10.0